Update google-federation.md
The language in this article as written is confusing. Microsoft distinguishes between Microsoft Accounts and Azure AD Accounts. I.e., a user@outlook.com account and a user@contoso.com account are very different things for the purposes of just about any technet article. For the purposes of adding Google as an identity provider and inviting guests to Teams, no such distinction exists between a personal@gmail.com Gmail account and a gsuiteuser@example.com Gmail account. This article would be more helpful to IT users supporting Teams users trying to add external users if it merely told people that they need to add Google as an identity provider (if they haven't already).